add GP_NO_WRAPPER option to game-performance to replace game-performance process with the game process

+++ scripts/game-performance.nix
@@ -6,11 +6,17 @@ export LD_PRELOAD=""
 export DXVK_FILTER_DEVICE_NAME="NVIDIA"
 export VKD3D_FILTER_DEVICE_NAME="NVIDIA"
 
-systemctl start scx.service
-cpuperf performance
-"$@"
-cpuperf powersave
-systemctl stop scx.service
+if [ "$GP_NO_WRAPPER" == "1" ]; then
+	systemctl start scx.service
+	cpuperf performance
+	exec "$@"
+else
+	systemctl start scx.service
+	cpuperf performance
+	"$@"
+	cpuperf powersave
+	systemctl stop scx.service
+fi
 	'';
 in [
 	script
